completed the youtube rag using the whisper model of open ai
